UI display added, modification to spawning, enemy respawn
UI display added made by Kieran.
Modified the listener on the beatmanager.cs so it will allow for variating spawn patterns if we wish for the future.
Addressed issue where the enemy would respawn oddly after the game has played for awhile. Anchor.cs was modified that corrected this issue, setparent null did the job.